docs(databricks): add guide for Databricks integration (#12326)
### Summary & Motivation
Here, we provide the basic template for using Dagster and Databricks,
using the resource abstraction.

Note that we do not yet provide a template for using a prebuilt asset
factory to generate these assets. This will be added later. For now, add
value by showing how the integration is set up in the first place.

### How I Tested These Changes
local, eyes