iteritems -> items (#12629)
### Summary & Motivation

Fixes https://github.com/dagster-io/dagster/issues/12604

`.iteritems` is [deprecated since 1.50
](https://pandas.pydata.org/docs/whatsnew/v1.5.0.html#other-deprecations)
and will be removed in a future version. Changing to `.items`.
(https://github.com/pandas-dev/pandas/pull/45321).

### How I Tested These Changes

I did not, but https://github.com/pandas-dev/pandas/pull/45321 appears
to indicate it's a safe change to make. If I can somehow install this
version of dagster on my machine, I'm happy to test using the code that
surfaced the issue, but I'm not sure how to do that.