Flatten overlapping time window subsets in order of priority (#12632)
This PR provides a helper method that accepts a list of
`TimeWindowPartitionsSubset`s and their statuses (failed or
materialized). Flattens these subsets into a list of range statuses
based on the priority of the statuses.

This helper function:
- calculates the overlap between the two subset's ranges
- removes the overlap from the lower priority subset's ranges

The hope is that this function in the future can be adapted to accept
`n` subsets and can flatten based on the priority of each subset.