[ui] Use code location perms on BackfillRow (#12648)
### Summary & Motivation

If a failed backfill has a partition set, use its code location to
determine permissions for resuming it. This is instead of using the
unscoped permission.

If the backfill has no partition set, this will fall back to the
unscoped permission value.

### How I Tested These Changes

I could use some help here. I'm not entirely sure how to test this
locally, because I'm not sure how to generate a failed backfill.