[dagit] Support displaying a “failed” partition status (#12698)
### Summary & Motivation

This PR adds support for @johannkm's new failed states to Dagit. This is
part 1 of 2, with updated designs for asset nodes still to come.

Here's a loom video of the changes, demoed on a variety of assets:

https://www.loom.com/share/cfd03a2cbf6d4846a4d8d6a332b6b5c4?from_recorder=1&focus_title=1
(There is a bug in the video where it shows "Failed(0)" in one spot,
fixed and updated in the screenshot below!)

There are two awkward bits that are worthy of follow-up:

- We don't yet have a "partially failed, partially materialized" or
"partially failed, partially missing" state for the bars, which will be
helpful for multi-dimensional assets. It introduces some fan out the way
it's currently built, because we store the state in an enum and it now
needs combinatorial values. I think this requires a larger refactor
because when we add "STALE" we'll have even more possible combinations.

- The "Only missing" feature in the materialize dialog doesn't play
super nicely with this because we actually don't know which of the
failed are missing + failed and which of the failed partitions were
previously materialized. __One immediate option would be to change this
to "Only missing and failed"__

### How I Tested These Changes

I edited the mock data by hand to embed some interesting failure states
and then updated the tests to match.

I also updated the mock data used by our Storybooks for asset health so
they render with some red bars:

<img width="1840" alt="image"
src="https://user-images.githubusercontent.com/1037212/222804201-1bf787ec-53d2-4813-b570-75a40d39816d.png">

<img width="1836" alt="image"
src="https://user-images.githubusercontent.com/1037212/222805603-d924d016-4780-4685-8034-8d1e74ed8675.png">

<img width="1836" alt="image"
src="https://user-images.githubusercontent.com/1037212/222807910-cbf41db3-05ce-44d6-8ad2-a415345b83a9.png">

---------

Co-authored-by: bengotow <bgotow@elementl.com>