Set default memory/CPU for new ECS tasks based on runtime platfform
Summary:
Windows task definitions need to be registered with at least 1024 CPU and 2048 memory, even if the cpu/memory are going to be later overridden by a task override / container override. Instead of hard-coding 256/512 for all tasks, check the runtime platform and change the default CPU/memory accordingly.

Additionally, if run_resources have been explicitly set, send those in as the task definition CPU/memory (This is harmless and is another fix for the problem above - it also probably makes it easier to understand what the actual resource limits are if you're inspecting the task).