[rename] LogicalVersion -> DataVersion (#12500)
### Summary & Motivation

Rename `LogicalVersion` -> `DataVersion` everywhere except for the GQL
schema and docs (this is done upstack).

The only remaing use of "logical" version is in the (internal) function
`compute_logical_data_version`, which is a rename of
`compute_logical_version`. "logical" was retained here as a qualifier to
"data version" to indicate the method of computation
(`compute_data_version` is too unspecified).

Note that this also updates the tags used to store logical versions and
input logical versions from:

- `dagster/logical_version` -> `dagster/data_version`
- `dagster/input_logical_version` -> `dagster/input_data_version`

This is important since we display the actual tags in dagit. All reading
of these tags should be done through `extract_data_version_from_entry`
and `extract_data_provenance_from_entry`, which will try to read both
the old and the new tags.

### How I Tested These Changes

Existing test suite + added a new unit test that accesses legacy tags.