Remove PartitionMetadataEntry (#12726)
### Summary & Motivation

Remove `PartitionMetadataEntry`. Consider this an RFC because I don’t
have full context here, but here is my understanding from a study of the
code:

- `PartitionMetadataEntry` is an experimental class that was added in
the past to support creation of metadata entries qualified by a single
partition.
- `PartitionMetadataEntry` is not a top-level export— it’s not part of
the public API. Unclear whether this is inadvertent.
- Dagit has no concept of a `PartitionMetadataEntry`.
- `PartitionMetadataEntry` is incompatible with our recommended
dictionary-based metadata API, since it does not allow the same metadata
label (key in the dict) to be used for multiple partititons.
- `PartitionMetadataEntry` is only used in a single test, which is
skipped.
- The only place the codebase has `PartitionMetadataEntry`-specific
logic is during the generation of `AssetMaterializations` associated
with an output. Here is what happens:
- Metadata entries yielded from the IO manager’s `handle_output` (this
runs before output-derived `AssetMaterializations` are generated) and
metadata entries passed on the `Output` object are combined.
- If any of these are `PartitionMetadataEntries`, they are resolved to
plain `MetadataEntries` and stored in a dict that maps partitions to
`MetadataEntry`.
- This table is consulted to obtain the `MetadataEntries` that will be
placed on each `AssetMaterialization` to be generated from an `Output`.
- The only way an output generates multiple `AssetMaterializations` is
if the run tags specify a `PartitionKeyRange`. In this case, for each
asset-associated-output in the run, one `AssetMaterialization` will be
generated for each partition in the range.
- Therefore in theory one could either (a) yield
`PartitionMetadataEntry` events from the IO manager while writing an
asset-associated output spanning multiple partititons; (b) return an
`Output` object in a materialization function with a number of
`PartitionMetadataEntry` objects associated to different partitions of
the output.

So, IIUC PartitionMetadataEntry was created to solve the problem of
providing individual-partitioned-scoped metadata entries for the case
whene a single run materializes multiple asset partitions. But:

It’s unclear to me whether materializing multiple partitions in one run
is something we even support now or if it’s an old abandoned API. The
[Partitions concept
page](https://docs.dagster.io/concepts/partitions-schedules-sensors/partitions#partitioned-asset-jobs)
doesn’t appear to mention it.

If my understanding above is right, I propose removing
PartitionMetadataEntry. It is incompatible with our recommended metadata
API and complicates the docstrings we present to users. In the event we
want to allow users to provide individual-partition-scoped metadata on
an Output object, we can allow passing a Mapping[str, MetadataDict] on
the output. We could introduce a similar API for yielding
partition-scoped metadata entries from handle_output, though it’s
unclear if that’s even something we want to support.

### How I Tested These Changes

BK