[ui] Affordance for copying asset key (#12829)
### Summary & Motivation

Add a small icon button next to the asset path breadcrumbs in the asset
page, allowing the user to copy the asset key path, joined by slashes.


https://user-images.githubusercontent.com/2823852/223870030-cc6cefd9-0fd7-42d3-ad66-a65fd44ac9c3.mov

### How I Tested These Changes

View an asset, hover on copy button and click it. Verify that copy
occurs properly, that the toaster appears, and that the icon changes
briefly.

Repeat with an asset that has many path keys, verify same.