Move from pull request template to contribution guide.
I am a huge fan of our docs style guidelines and making them publicly available.

However, I'd also like to advocate for keeping our pull request template as lightweight as possible, to minimize cognitive load when submitting a PR. It used to have a lot more in it, and @alangenfeld made a change to trim it down to the bare essentials: https://github.com/dagster-io/dagster/pull/7538.

This change adds the style guidelines to the top of the docs section of the contribution and removes it from the PR template.