[ui] Refactor permissions objects (#12903)
## Summary & Motivation

Work toward repairing the Permissions footgun, in which it's too easy to
check the `PermissionResult` object for truthiness instead of looking at
its `enabled` value.

Return a `PermissionsState` object from `useUnscopedPermissions` and
`usePermissionsForLocation`. This includes a `permissions` map
(booleans), a `disabledReasons` map (strings), and a `loading` boolean.

This means that for a given permission, the "enabled" state is separate
from its `disabledReason`, and truthiness checks are simple boolean
checks.

I have a temporary shim for Cloud types that I'll clean up right away.

## How I Tested These Changes

TS, lint, jest. Load Cloud and view a job with the "Viewer" role, verify
that the launchpad is disabled.