[dagit] Allow the lineage tab to appear before the asset definition has loaded (#12191) (#12814)
### Summary & Motivation

This PR greatly reduces flicker observed in #12191. Instead of waiting
for the asset definition to load and then loading the content of the
current tab, this PR moves the loading state determination down into
each tab's render method and refactors the lineage tab so that it does
not require the loaded definition.

Impact: when clicking between assets on the asset lineage tab, the
lineage UI very briefly flickers as it moves between assets rather than
taking you through two spinner / loading states.

Caveat: There's still a short flicker. When you navigate to an asset
that you haven't visited before, the `AssetsCatalogRoot` requests
assetOrError to decide whether to display the catalog table or the asset
details view. This is a nasty side effect of our decision to make
`assets/123` show either the list of assets at that path OR the asset
detail page, depending on whether 123 is the name of an asset. I think
we should revisit this separately.

### How I Tested These Changes

I slowed down requests in my browser using the network tab and verified
that there is now just a single brief loading state. I also verified
that going to each tab and Cmd-R reloading does not cause any issues
(eg: when data is missing / being fetched for the first time and you are
viewing each tab)

Co-authored-by: bengotow <bgotow@elementl.com>