Fix valid start time check for time partitions (#12944)
Reported by a user:
https://dagster.slack.com/archives/C01U954MEER/p1678304642514719

Fixes an error where the first partition in the partition health bar
always displays as unmaterialized despite a materialization existing for
this partition. This error occurs when the running OS timezone is ahead
of the partitions def timezone (by default UTC).

The cause of this issue was a partition key validation check that relied
on a `datetime.strptime` object which is not timezone aware (so defaults
to the running OS timezone). Because the external time window partitions
def creates a start time object that is timezone aware, certain valid
partition keys would be considered invalid as they were set to an
earlier timezone and thus were considered to be "before" the valid start
time.