[ruff] Enable D415 (docstring period) (#12892)
## Summary & Motivation

Turns on Ruff `D415` docstring header punctuation (required by
Google-style docstrings). This was supposed to be turned on in the past
(see https://github.com/dagster-io/dagster/pull/11001) but I had not
finished correcting the old violations in part because IIRC the autofix
at the time was bugged for some cases. Autofix works well now so we can
have this on, ensuring docstrings shown in docs and LSP service are
well-formatted, with little-to-no additional effort required for devs.

In a few cases I had to add a suitable header docstring if the existing
one wasn't a summary.

---

## How I Tested These Changes

BK