[dagit] Preserve key order when filtering on the asset partitions page (#12952)
## Summary & Motivation

This is a small patch for an issue I noticed today where filtering the
asset partitions page can result in partition keys appearing out of
order.

There were two errors that caused this:

- The fast-path case was not being matched because the default "all
three selected" value was in a different order than expected. (I think
this broke when we added FAILED)

- The code that builds the "success + missing" or "failure + missing"
set was concatenating the success keys with the missing keys. This is
the fastest way to return the right set of keys, but ignores the
ordering. I think the best way to do this and preserve ordering is to
walk the keys array exactly once. Alternatively we could create ranges
for the missing segments and run them through the same logic we already
have, but it's not something we do elsewhere and violates the TS types.

<img width="564" alt="Screen Shot 2023-03-14 at 6 35 05 PM"
src="https://user-images.githubusercontent.com/1037212/225166498-a5c42297-7685-41f7-928c-03560c035bc5.png">


## How I Tested These Changes

- I tested this manually by fiddling with the filters on the asset
details page and verifying it shows the correct partitions in the
sidebar in the correct ordering. I think I can get this
`dimensionKeysInSelection` method under test in a follow-up PR, but
we'll need to pull it out and change the params.

Co-authored-by: bengotow <bgotow@elementl.com>