[ui] Use a multi-select component for log filters (#13004)
## Summary & Motivation

Resolves #12741.

On the Run page:

- Replace the log level "switch-tags" with a single select component
that allows choosing multiple log levels. In the menu of the component,
show the number of logs for that level, as we've done with the
switch-tags.
- When changing log levels, the URL querystring will update as before.
However, we will also now persist the enabled log levels so that these
will be the defaults the next time you visit a Run.

<img width="219" alt="Screenshot 2023-03-17 at 8 37 32 AM"
src="https://user-images.githubusercontent.com/2823852/225926773-56d14abd-460d-4be3-87d0-235e6c54bcd8.png">


https://user-images.githubusercontent.com/2823852/225926753-ceff0068-aa22-45ac-b393-8e69f02b8faf.mov

## How I Tested These Changes

Storybook example for LogFilterSelect.

View a run in the app. Change log levels, verify that the log table and
querystring update accordingly. Reload the page with no query
parameters, verify that my log level selections persist. Navigate to
another Run, verify same.