Only create instance when necessary in GRPC server in partitions endpoints (#12678)
Fixes https://github.com/dagster-io/dagster/issues/12440

The gRPC server attempts to instantiate the dagster instance for methods
such as fetching partition tags / config / execution data. This causes
these endpoints to fail when instantiating the instance if the gRPC
server does not have access to the instance (for example if its deployed
as a container). This PR fixes the issue by checking if the requested
partitions def is dynamic, only instantiating the instance if it is.