[Multipartitions w/ dynamic dimension 2/2] Cache dynamic partitions status (#12897)
Part 1: https://github.com/dagster-io/dagster/pull/12894

This PR enables caching dynamic partitioned asset statuses in the asset
status cache value. Since we now allow dynamic partitions defs to be
dimensions of multipartitions defs and we cache multipartitions
statuses, this PR brings the dynamic partitions status caching up to
parity.

One side effect is that the serializable_unique_identifier we use to
check whether a partitions definition has remained the same or not must
be changed to a function that accepts a DynamicPartitionsStore, in order
to build an identifier containing the dynamic partition keys. This
enables rebuilding the partitions status cache when the partitions
definition has been mutated.

Performance is expected to remain the same with this change. Dynamic
partitions are serialized in default partitions subsets, which contain
the set of selected partition keys, so querying the instance is not
necessary to fetch the partition keys. We have a caching dynamic
partitions loader that exists on all asset nodes, so when querying for
partition statuses / partition keys for a set of assets we only query
the instance once for the partition keys of a given dynamic partitions
definition and return the cached value thereafter.