[refactor] delete Materialization (#13030)
## Summary & Motivation

Delete `Materialization` from `dagster._legacy`.

This turned out to be extremely simple-- just needed to add
`Materialization` to `old_storage_names` of `AssetMaterialization`. This
works because `Materialization` has one extra field (`label`) that is
dropped on deserialization, and `AssetMaterialization` has an extra
field `tags` that gets a default on deserialization. Otherwise the two
classes are the same.

## How I Tested These Changes

Added test `test_load_old_materialization` to ensure an old serialized
`Materialization` loads as an `AssetMaterialization`.