Greatly reduce the number of gRPC calls when doing large asset backfills
Summary:
By caching the externalpipeline and externalexecutionplan when they are identical, we can greatly reduce the amount of back and forth that we have to do with the gRPC server. When I run a 2500-run backill of hourly_partitioned_asset, this moves it from 5000 gRPC calls to 2.

Test Plan: Run the above backfill locally