Add dynamic partitions name resolver to dimension type (#13070)
Currently the Dagit view to add a partition to a multipartitions def
containing a dynamic dimension is broken and throws a "you do not have
permission" error. The cause of this issue is that the name of the
multipartitions def (which is `null`) is being passed into the add
partitions mutation, so no matching partitions defs are found.

This PR adds a `dynamicPartitionsName` resolver to the partition
dimension gql object, to fetch the name of nested dynamic partitions
defs within multipartitions defs. This resolver is not to be confused
with the `name` resolver, which represents the dimension name (though
longer term we should refactor these objects to be cleaner).

In addition, modifies the permission check to also detect when a
matching partitions def is found that is a dimension a multipartitions
def.