Fix submitting backfills synchronously from graphql
Summary:
This param is only used under tests, but in the case where backfills are being submitted from graphql we should use the same request context for the length of the request (instead of passing in the process context like we do in the daemon). Add a lambda so that we can vary the behavior across the daemon and graphql.

Test Plan: Internal BK should pass now instead of erroring when trying to create a request context in cloud