ExitStack.pop_all -> close (#13050)
The current call sites appear to believe this is equivalent to close
even though it is not quite. Clearer to use the correct method.
 
> pop_all()
> Transfers the callback stack to a fresh ExitStack instance and returns
it. No callbacks are invoked by this operation - instead, they will now
be invoked when the new stack is closed (either explicitly or implicitly
at the end of a with statement).


## How I Tested These Changes

bk