celery-k8s executor: handle stdout that's valid json but not a dagster event (#13143)
Celery-k8s step workers write json serialized events to stdout. The
celery workers parse them. It seems like if you write a line to stdout
that parses as json but isn't an event, we fail with a
DeserializationError:
https://dagster.slack.com/archives/C01U954MEER/p1679635251354799

https://github.com/dagster-io/dagster/pull/7665 added 

```
except JSONDecodeError:
            pass
```

in order to skip lines that don't parse as valid json. We can add
DeserializationError to the errors to skip, so we just ignore lines that
aren't events.

Of course ideally we should switch to reading events from the db but
that's a larger change. I don't think there's too much risk of true
errors getting swallowed by this?

---

Side note:

I'm not sure what the 

```
        except check.CheckError:
            pass
```

catches. Is it possible that before the
`deserialize_json_to_dagster_namedtuple`, this would raise a CheckError
instead of a DeserializationError? Doesn't look like it from some quick
browsing but I am a bit surprised this hasn't come up before. Maybe bc
it's rare to write directly to stdout?

---

Testing the test here: https://github.com/dagster-io/dagster/pull/13144