[docs][tutorial-revamp] Adding a section for scheduling to the tutorial (#13101)
## Summary & Motivation

This PR adds a page on schedules. It touches jobs, code locations, and
schedules.

A non-tutorial change is that it also moves the code around a bit for
the template used for `dagster scaffold`. This is almost purely
stylistic, so I can punt it or move it to a separate PR if asked.

Ended up taking Sandy's guidance and moving the part about adding
definitions to a more just-in-time place. Now, definitions are mentioned
lightly at the point where the reader will have to add a job to the
`Definitions` object.

## How I Tested These Changes


https://dagster-git-docs-tutorialscheduling-elementl.vercel.app/tutorial/scheduling-your-pipeline

---------

Co-authored-by: Erin Cochran <erin.k.cochran@gmail.com>