Add support for more asset tags (#13153)
## Summary & Motivation
This adds support for ~30 new asset/op tags and adds a new `reversed`
property that turns the tag text black to better support icons with low
contrast colors (like yellow and orange). I also removed the light gray
background on tags in the list view to give them bit higher contrast.

How I created this list of icons:
- looked at the list of technologies our users noted in the latest user
survey
- looked at the list of integrations we support
- added some common ML libraries and other data tools

This list is not exhaustive, we will continue to add to it but this PR
gives us a much more robust list of icons that can be used in the asset
graph and possibly leveraged in other parts of the UI in the future.


![image](https://user-images.githubusercontent.com/2798333/227589357-aef08d0a-3513-462c-84ae-566f456873b7.png)


## How I Tested These Changes
Created a bunch of assets, added a tag for each one. Reviewed dagit.
https://www.loom.com/share/438a6e37c9284d5a9acd28f8b8f7073f

---------

Co-authored-by: bengotow <bgotow@elementl.com>