MultiPartitionMapping (#12950)
Requested by multiple users, primarily for the use case of defining
multipartitions dependencies for different time intervals:
https://github.com/dagster-io/dagster/issues/12905

Introduces a new `MultiPartitionsMapping` partition mapping that allows
for defining partition mapping relationships per-dimension. For example,
the below partition mapping would map `a|2023-01-01` in `weekly_abc`
would map to `1|2023-01-01`... `1|2023-01-07`.
```
weekly_abc = MultiPartitionsDefinition(
    {
        "abc": StaticPartitionsDefinition(["a", "b", "c"]),
        "weekly": WeeklyPartitionsDefinition("2023-01-01"),
    }
)
daily_123 = MultiPartitionsDefinition(
    {
        "123": StaticPartitionsDefinition(["1", "2", "3"]),
        "daily": DailyPartitionsDefinition("2023-01-01"),
    }
)

MultiPartitionsMapping(
    {
      "abc": DimensionMapping(
          downstream_dimension_name="123",
          partition_mapping=StaticPartitionMapping({"a": "1", "b": "2", "c": "3"}),
      ),
      "weekly": DimensionMapping(
            downstream_dimension_name="daily",
            partition_mapping=TimeWindowPartitionMapping(),
       )
    }
)
```
The `MultiPartitionsMapping` object checks that the selected dimensions
in the dimension mapping objects are existent, relying on the partition
mapping per-dimension to perform type checks and calculate the mapped
partitions. Each dimension in the upstream partitions definition must be
mapped 1:1 with a dimension in the downstream partitions definition.

In the situation where a user only wants to define a dependency
relationship between one dimension of the upstream and downstream
multipartitions defs, they could do something like the below example. In
this case, `a|2023-01-01` would map to `1|2023-01-01`, `2|2023-01-01`,
and `3|2023-01-01`.
```
MultiPartitionsDefinition(
    {
        "abc": StaticPartitionsDefinition(["a", "b", "c"]),
        "daily": DailyPartitionsDefinition("2023-01-01"),
    }
),
MultiPartitionsDefinition(
    {
        "123": StaticPartitionsDefinition(["1", "2", "3"]),
        "daily": DailyPartitionsDefinition("2023-01-01"),
    }
),
MultiPartitionsMapping(
    {
        "abc": DimensionMapping(
            "123",
            AllPartitionMapping(),
        ),
        "daily": DimensionMapping(
            "daily",
            IdentityPartitionMapping(),
        ),
    }
)
```
I'm very open to API changes. My intent here was to make this partition
mapping generic, so it can handle any number of dimensions in the
future.